fix bug with temp dupe item
